some more refactoring
