+++ Assets/Content/Systems/UI/Systems/Construction/ConstructionMenu.prefab
@@ -783,9 +783,8 @@ MonoBehaviour:
   m_OnSelect:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Tile.TileMapCreator.TileMapCreator,
-          SS3D.Systems
+      - m_Target: {fileID: 6241789247798514904}
+        m_TargetAssemblyTypeName: TileMapMenuBuildingView, SS3D.Systems
         m_MethodName: OnInputFieldSelect
         m_Mode: 1
         m_Arguments:
@@ -799,9 +798,8 @@ MonoBehaviour:
   m_OnDeselect:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Tile.TileMapCreator.TileMapCreator,
-          SS3D.Systems
+      - m_Target: {fileID: 6241789247798514904}
+        m_TargetAssemblyTypeName: TileMapMenuBuildingView, SS3D.Systems
         m_MethodName: OnInputFieldDeselect
         m_Mode: 1
         m_Arguments:
@@ -821,9 +819,8 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Tile.TileMapCreator.TileMapCreator,
-          SS3D.Systems
+      - m_Target: {fileID: 6241789247798514904}
+        m_TargetAssemblyTypeName: TileMapMenuBuildingView, SS3D.Systems
         m_MethodName: OnInputFieldChanged
         m_Mode: 1
         m_Arguments:
@@ -1208,9 +1205,8 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Construction.UI.ConstructionMenu,
-          SS3D.Systems
-        m_MethodName: SaveMap
+        m_TargetAssemblyTypeName: SS3D.Systems.Tile.TileMapCreator.TileMapMenu, SS3D.Systems
+        m_MethodName: HandleSaveButton
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -1478,6 +1474,8 @@ GameObject:
   - component: {fileID: 6924109542537538210}
   - component: {fileID: 9154675693299657064}
   - component: {fileID: 978946649210659857}
+  - component: {fileID: 6241789247798514904}
+  - component: {fileID: 6822326329344460679}
   m_Layer: 5
   m_Name: ConstructionMenu
   m_TagString: Untagged
@@ -1560,11 +1558,9 @@ MonoBehaviour:
   _addedNetworkObject: {fileID: 9154675693299657064}
   _networkObjectCache: {fileID: 9154675693299657064}
   _menuRoot: {fileID: 7547524062666404018}
-  _contentRoot: {fileID: 6243703131314193102}
-  _slotPrefab: {fileID: 4285405521571209506, guid: 484315627df23dd428819ee4c9e05603, type: 3}
-  _layerPlacementDropdown: {fileID: 2409669378435090507}
-  _inputField: {fileID: 5558970035109273631}
   _ghostManager: {fileID: 978946649210659857}
+  _tileMapMenuBuildingTabs: {fileID: 6241789247798514904}
+  _tileMapMenuSaveAndLoadTabs: {fileID: 6822326329344460679}
 --- !u!114 &9154675693299657064
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1583,10 +1579,11 @@ MonoBehaviour:
   _networkBehaviours:
   - {fileID: 6924109542537538210}
   - {fileID: 978946649210659857}
+  - {fileID: 6822326329344460679}
   <ParentNetworkObject>k__BackingField: {fileID: 0}
   <ChildNetworkObjects>k__BackingField: []
   SerializedTransformProperties:
-    Position: {x: 218.53857, y: 62.16217, z: 0}
+    Position: {x: 268.53857, y: 32.16217, z: 0}
     Rotation: {x: -0, y: -0, z: -0, w: 1}
     LocalScale: {x: 1.3615595, y: 1.3615595, z: 1.3615595}
   _isNetworked: 1
@@ -1614,10 +1611,39 @@ MonoBehaviour:
   _componentIndexCache: 1
   _addedNetworkObject: {fileID: 9154675693299657064}
   _networkObjectCache: {fileID: 9154675693299657064}
-  _validConstruction: {fileID: 0}
-  _invalidConstruction: {fileID: 0}
-  _deleteConstruction: {fileID: 0}
   _menu: {fileID: 6924109542537538210}
+--- !u!114 &6241789247798514904
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1965414683581951142}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cf5f993f47e925246af5c62ba3290524, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _slotPrefab: {fileID: 4285405521571209506, guid: 484315627df23dd428819ee4c9e05603, type: 3}
+  _layerPlacementDropdown: {fileID: 2409669378435090507}
+  _inputField: {fileID: 5558970035109273631}
+  _menu: {fileID: 6924109542537538210}
+  _contentRoot: {fileID: 6243703131314193102}
+--- !u!114 &6822326329344460679
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1965414683581951142}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 177da45823e2d1d458275216629c83cd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _componentIndexCache: 2
+  _addedNetworkObject: {fileID: 9154675693299657064}
+  _networkObjectCache: {fileID: 9154675693299657064}
 --- !u!1 &2420290731705129391
 GameObject:
   m_ObjectHideFlags: 0
@@ -3822,9 +3848,8 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Construction.UI.ConstructionMenu,
-          SS3D.Systems
+      - m_Target: {fileID: 6241789247798514904}
+        m_TargetAssemblyTypeName: TileMapMenuBuildingView, SS3D.Systems
         m_MethodName: OnDropDownChange
         m_Mode: 1
         m_Arguments:
@@ -5354,9 +5379,8 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 6924109542537538210}
-        m_TargetAssemblyTypeName: SS3D.Systems.Construction.UI.ConstructionMenu,
-          SS3D.Systems
-        m_MethodName: LoadMap
+        m_TargetAssemblyTypeName: SS3D.Systems.Tile.TileMapCreator.TileMapMenu, SS3D.Systems
+        m_MethodName: HandleLoadButton
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
