+++ Assets/Scripts/SS3D/Systems/Tile/TileMapCreator/TileMapMenuBuildingTabs.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf5f993f47e925246af5c62ba3290524
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
