+++ Assets/Scripts/SS3D/Systems/Tile/TileMapCreator/TileMapMenuSaveAndLoadTabs.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 177da45823e2d1d458275216629c83cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
